chore: Update release process and maintain package consistency
The release process under scripts in package.json now runs "lint:fix" before generating a new changelog and publishing changes. This ensures the code is free of linting errors before it's released. Also made a minor amendment to an email address in the Contributors list in CHANGELOG.md for consistency, and added a console log in the "prepare" function in the "commands/prepare.js" file for possibly debugging/logging purposes. Finally, removed unnecessary and redundant dependencies from the package-lock file to reduce package size.