fix!: Update `for` loops to `forEach` for object iteration
Updated instances of 'for' loops attempting to iterate over objects in `update.cjs` and `prepare.cjs` files to use 'forEach', which is better suited for this purpose. An incorrect passing of keys in the 'for' loop in `update.cjs` was also corrected. These changes ensure accurate key-value pair iteration over `nameMapping` and `prodObject` objects.