feat: add new CLI command to register microfrontend
A new command(`add`) is introduced to the CLI. This `add` command is used for registering a new microfrontend in our application. This is done by parsing the production URL to parts and then storing that information in the package.json file as well as in the `microFrontendURL.js` configuration file. The import and use of the `add` command was done in our CLI file (cli.js) and the necessary registration logic in the new files `commands/add.js` and `tasks/add.cjs`.