feat: Update the URL template in the add Task; Port is now configurable
The URL template in the "add" Task was updated to make the port configurable. Previously, the port was hardcoded to 3000 for localhost. Now, it's set as a new "port" property in the configuration object based on the value passed to the add function. This allows for more flexibility when configuring and using the task in different environments. In addition, the "add" command parameters were updated to include an optional "local-port" to facilitate this change.